refactor(client): abstract RawTextArea for variable input
