+++ packages/core/client/src/schema-component/antd/variable/Variable.tsx
@@ -3,6 +3,7 @@ import { connect, mapReadPretty } from '@formily/react';
 import { IField } from '../../../collection-manager';
 import { Input } from './Input';
 import { JSONInput } from './JSONInput';
+import { RawTextArea } from './RawTextArea';
 import { TextArea } from './TextArea';
 
 export function Variable() {
@@ -12,6 +13,8 @@ export function Variable() {
 Variable.Input = connect(Input);
 
 Variable.TextArea = connect(TextArea, mapReadPretty(TextArea.ReadPretty));
+
+Variable.RawTextArea = connect(RawTextArea);
 
 Variable.JSON = connect(JSONInput);
